+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes relevant to the users of python-tcod are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/) since version `2.0.0`.
 
 ## [Unreleased]
+### Fixed
+- Fixed the parsing of SDL 2.0.22 headers.  Specifically `SDL_FLT_EPSILON`.
 
 ## [13.6.2] - 2022-05-02
 ### Fixed